Remove more unnecessary v2 mentions
